requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.2
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
urllib3==2.5.0
Werkzeug==3.1.3
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.116.2
Flask==3.1.2
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3